Feito Relacionamento entre pessoa e familia, e adequado os testes.
