+++ src/test/java/com/InteligenciaEletricaAPI/endereco/EnderecoTests.java
@@ -1,4 +1,4 @@
-package com.InteligenciaEletricaAPI;
+package com.InteligenciaEletricaAPI.endereco;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
