+++ src/test/java/com/InteligenciaEletricaAPI/equipamento/EquipamentoTests.java
@@ -1,4 +1,4 @@
-package com.InteligenciaEletricaAPI;
+package com.InteligenciaEletricaAPI.equipamento;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
