+++ src/test/java/com/InteligenciaEletricaAPI/familia/FamiliaTests.java
@@ -1,4 +1,4 @@
-package com.InteligenciaEletricaAPI;
+package com.InteligenciaEletricaAPI.familia;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
